lab stations in editable form adjust to chosen lab
When a lab is chosen in the editable form the lab stations change
according to the lab that is chosen.

Also in this commit: Disiplinary meeting field in editable form can be
cleared. When saved if this field is empty then meetingDate in the model
will be updated to 'Pending'