each comment is now ordered by a unix time stamp
The time stamp stamp is made from current time whne comment is created. It
updates using the input time and date whenever the comment is edited and
saved.

Moment.js is used to do some conversions. It is probably not necessary
though with the realization that most of the functionality here can be
done fairly easily with the native JS Date object. Also Angular has it's
own filters for dates, which can probably be made use of in the
controllers.