name is searched for when typing in ID
When user types in a number ID all the incidents are searched for that
number. When one is found it places the name in that incident into the
name field.

This needs to be changed to search for Lab-ID instead of just the
number.

Also hard codes in the Pool in "Student worker's lab" options and
re-aranges some of the input fields.