incident controller now uses relative last url
This commit fixed the issue with the incident controller (currently called
just Ctrl) so that when an icident is deleted it will return to the
previous view location no matter which incident view is being looked at.
The last location is set in $rootScope then called from there when needed.