basic interface working - empty messages showing
The interface should be basically working correctly.

In Open Incident View when empty a message shows.

In the Report View, if incidents have all been moved or deleted for a
report the report also shows a message stating no report exist.

This may all still be confusing and does not make for the most intuitive
interface, but it should all be working in general.