basic login working with temp chrome extesion hack
Special use case: A particular chrome extesnios was not
working with Angular. The extension replaces the passsword in the DOM
right after the user types it in but Angular is not updating it in the model.
This commit hacks around that by just direclty pulling the value from
the DOM. This is a temp fix though as it is very bad style.