fix for protocall-less URL by specifying protocol
The sockets seemed to be connecting as browser was not throwing errors
but would throw them  when the server was turned off.
No messages were arriving or being sent though. It appears the
protocall-less URL was not working. Specifying the protocol fixes
things. This is an https only site so shouldn't really matter to
keep it this way.