set up simplified server to run on heroku
This commit takes the original app.js and renames it to app2.js so
it's easily accessible.

Also the easyIn function is added to auth.js to atutomatically
authenticate any logins for demo purposes for now.

This should get the app up and running on heroku while all the other
functionality can be slowly built back into the app.